+++ vendor-prefixed/guzzlehttp/guzzle/src/Cookie/CookieJarInterface.php
@@ -8,8 +8,8 @@
 
 namespace WPGraphQL\Login\Vendor\GuzzleHttp\Cookie;
 
-use Psr\Http\Message\RequestInterface;
-use Psr\Http\Message\ResponseInterface;
+use WPGraphQL\Login\Vendor\Psr\Http\Message\RequestInterface;
+use WPGraphQL\Login\Vendor\Psr\Http\Message\ResponseInterface;
 
 /**
  * Stores HTTP cookies.