+++ vendor-prefixed/guzzlehttp/guzzle/src/Exception/GuzzleException.php
@@ -8,7 +8,7 @@
 
 namespace WPGraphQL\Login\Vendor\GuzzleHttp\Exception;
 
-use Psr\Http\Client\ClientExceptionInterface;
+use WPGraphQL\Login\Vendor\Psr\Http\Client\ClientExceptionInterface;
 
 interface GuzzleException extends ClientExceptionInterface
 {