+++ vendor-prefixed/league/oauth2-linkedin/src/Provider/LinkedIn.php
@@ -16,7 +16,7 @@ use WPGraphQL\Login\Vendor\League\OAuth2\Client\Provider\Exception\LinkedInAcces
 use WPGraphQL\Login\Vendor\League\OAuth2\Client\Token\AccessToken;
 use WPGraphQL\Login\Vendor\League\OAuth2\Client\Token\LinkedInAccessToken;
 use WPGraphQL\Login\Vendor\League\OAuth2\Client\Tool\BearerAuthorizationTrait;
-use Psr\Http\Message\ResponseInterface;
+use WPGraphQL\Login\Vendor\Psr\Http\Message\ResponseInterface;
 
 class LinkedIn extends AbstractProvider
 {